Add Hexagon V75 CMake build for YNNPACK
- This commit is an addition to the branch 'test_890235712'
- cmake/hexagon.toolchain.v75 + scripts/build-qurt-v75.sh: toolchain
  and driver script for the QuRT v75 cross build.
- pthreadpool: mirror XNNPACK's pattern. Add cmake/DownloadPThreadPool
  .cmake and a pthreadpool block in CMakeLists.txt so QuRT builds
  resolve <pthreadpool.h>.
- Patch slinky on hexagon only: cmake/slinky_hexagon_int_disambiguation
  .patch adds explicit int overloads guarded by __hexagon__, since
  hexagon-clang can't disambiguate `int` -> `slinky::expr` against the
  int32_t/int64_t/size_t constructors. Applied via `patch -N` so
  re-runs are idempotent.
- Switch gtest discovery to PRE_TEST mode: the default POST_BUILD
  discovery runs each test binary at build time on hexagon-sim and
  times out before QuRT boots.
- consistent_arithmetic_test.cc: static_cast<float> before
  std::isfinite -- hexagon-clang couldn't resolve the overload on the
  buffer accessor's return type.

+++ ynnpack/kernels/dot/consistent_arithmetic_test.cc
@@ -139,8 +139,8 @@ void TestMatMul(AT, BT, CT, size_t k) {
       int finite = 0;
       for (size_t i = 0; i < m; ++i) {
         for (size_t j = 0; j < n; ++j) {
-          bool c_finite = std::isfinite(c(i, j));
-          bool kernel_c_finite = std::isfinite(kernel_c(i, j));
+          bool c_finite = std::isfinite(static_cast<float>(c(i, j)));
+          bool kernel_c_finite = std::isfinite(static_cast<float>(kernel_c(i, j)));
           if (c_finite && kernel_c_finite) {
             ASSERT_EQ(c(i, j), kernel_c(i, j));
             finite++;
